Revert "feat(core): widen length path to 32-bit, lifting the 64KB SDU cap"
The 32-bit widening was disproportionate: real UDS reprogramming chunks
payloads via TransferData (each block bounded by the negotiated
maxNumberOfBlockLength and the ECU buffer), so a single >64KB ISO-TP SDU
is not a real use case, and the change forced a breaking API churn across
every length-bearing callback for no practical gain.

Keep the meaningful part (escape FirstFrame + FlowControl WAIT/OVFLW).
The escape format still serves 4096..65535-byte SDUs and answers
FC.OVFLW above the receive buffer, with no API break.

This reverts commit 39c601a748a8b4b058bc02bef35eb6a9a18e089d.

+++ examples/bare_metal/main.c
@@ -21,7 +21,7 @@ static uint32_t get_system_time_ms(void)
     return 0;
 }
 
-static int can_send_frame(const uint8_t *data, uint32_t len)
+static int can_send_frame(const uint8_t *data, uint16_t len)
 {
     /* TODO: Write to CAN controller registers */
     return 0;
@@ -33,7 +33,7 @@ static uds_config_t cfg;
 static uint8_t rx_buffer[4096];
 static uint8_t tx_buffer[4096];
 
-static int tp_send(struct uds_ctx *ctx, const uint8_t *data, uint32_t len)
+static int tp_send(struct uds_ctx *ctx, const uint8_t *data, uint16_t len)
 {
     /* For Bare Metal without ISO-TP stack, use valid TP implementation */
     /* If using raw CAN, we need uds_tp_isotp.c (not shown here to keep simple) */
